// src/modules/pdf/models/extracted-product-item.model.ts
export interface ExtractedProductItem {
  productCode: string
  description: string
  orderNumber: string
  cean: string
  ncmSh: string
  cst: string
  cfop: string
  unit: string
  quantity: number
  unitValue: number
  totalValue: number
  icmsBaseValue: number
  icmsValue: number
  ipiValue: number
  icmsRate: number
  ipiRate: number
}